refactor: extract main's code to run
