+++ crd-from-oas/pkg/generator/generator_test.go
@@ -536,3 +536,59 @@ func TestCleanSingleVariantName(t *testing.T) {
 		})
 	}
 }
+
+func TestParseSDKTypePath(t *testing.T) {
+	tests := []struct {
+		name       string
+		input      string
+		wantImport string
+		wantType   string
+		wantErr    bool
+	}{
+		{
+			name:       "valid SDK type path",
+			input:      "github.com/Kong/sdk-konnect-go/models/components.CreatePortal",
+			wantImport: "github.com/Kong/sdk-konnect-go/models/components",
+			wantType:   "CreatePortal",
+		},
+		{
+			name:       "valid path with nested packages",
+			input:      "github.com/Kong/sdk-konnect-go/models/operations.ListPortals",
+			wantImport: "github.com/Kong/sdk-konnect-go/models/operations",
+			wantType:   "ListPortals",
+		},
+		{
+			name:    "no dot separator",
+			input:   "noDotAtAll",
+			wantErr: true,
+		},
+		{
+			name:    "leading dot",
+			input:   ".CreatePortal",
+			wantErr: true,
+		},
+		{
+			name:    "trailing dot",
+			input:   "github.com/Kong/sdk-konnect-go/models/components.",
+			wantErr: true,
+		},
+		{
+			name:    "empty string",
+			input:   "",
+			wantErr: true,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			importPath, typeName, err := ParseSDKTypePath(tc.input)
+			if tc.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, tc.wantImport, importPath)
+			assert.Equal(t, tc.wantType, typeName)
+		})
+	}
+}
